API

SteadyWaves.SteadyWaves

SteadyWaves.Params.CurrentCriterion

SteadyWaves.Params.DepthType

SteadyWaves.Params.ElevationType

SteadyWaves.Params.ParameterCriterion

SteadyWaves.Params.WaveType

SteadyWaves.Physics.PhysicsStruct

SteadyWaves.Surface.SurfaceStruct

SteadyWaves.Velocity.VelocityStruct

SteadyWaves.Wave.WaveStruct

SteadyWaves.Linear.linear_solution

SteadyWaves.Output.elevation

SteadyWaves.Output.horizontal_velocity

SteadyWaves.Output.pressure

SteadyWaves.Output.surface_tension

SteadyWaves.Output.vertical_velocity

SteadyWaves.Output.wave_height

SteadyWaves.Output.wave_number

SteadyWaves.Output.wave_period

SteadyWaves.Output.wave_period

SteadyWaves.Output.wave_power

SteadyWaves.Output.wavelength

SteadyWaves.Output.wavelength

SteadyWaves.Physics.validate_constants

SteadyWaves.Physics.validate_parameters

SteadyWaves.Shoaling.topo_approx

SteadyWaves.Shoaling.update_depth_fourier_approx

SteadyWaves.Steady.fourier_approx

SteadyWaves.SteadyWaves — Module

SteadyWaves is an implementation of Rienecker and Fenton (1981) Fourier Approximation Method to steady, periodic, nonlinear waves propagating in water of constant depth.

SteadyWaves.Velocity.VelocityStruct — Type

Structure with velocity field properties:

x: horizontal velocity

z: vertical velocity

psi: value of stream function

b: dimensionless amplitudes of stream function

SteadyWaves.Surface.SurfaceStruct — Type

Structure with properties of the free surface:

point: properties at control points

keypoints: free surface at keypoints

min: minimum of the free surface

max: maximum of the free surface

avg: average of the free surface

e_p: potential energy of the wave

a: amplitudes of wave frequencies

z: properties of the free surface at any x

SteadyWaves.Wave.WaveStruct — Type

Structure with wave properties:

eta: structure with properties related to free surface

v: structure with properties related to velocity

D: depth

C: celerity

R: wave-related parameter

H: wave height

U: average velocity

Q: flow-related parameter

N: number of control points along a wave

L: wavelength

T: wave period

F: wave flux

P: wave pressure function P(x, z)

sigma: surface tension coefficient

raw: internal representesion of equation system variables

SteadyWaves.Physics.PhysicsStruct — Type

Structure with physical constants:

g - gravitational acceleration [g]=m/s^2

rho - density of water [ρ]=kg/m^3

sigma - surface tension coefficient [σ]=kg/s^2

SteadyWaves.Physics.validate_constants — Method

Validate physical constants:

g,rho,sigma are positive

SteadyWaves.Physics.validate_parameters — Method

Validate parameters of wave:

H is positive

d is positive

Exactly one of L and T is positive and not nothing

SteadyWaves.Output.elevation — Method

elevation(w::WaveStruct, kx)

Calculate free surface elevation from solution w at coordinate kx or x.

If w is dimensional provide dimensional x. If not provide k*x.

SteadyWaves.Output.horizontal_velocity — Method

horizontal_velocity(w, kx, kz)

Calculate horizontal velocity from solution w at coordinates (kx, kz) or (x, z).

If w is dimensional provide dimensional (x, z). If not provide (k*x, k*z).

SteadyWaves.Output.pressure — Method

pressure(w, kx, kz)

Calculate pressure from solution w at coordinates (kx, kz) or (x, z).

If w is dimensional provide dimensional (x, z). If not provide (k*x, k*z).

SteadyWaves.Output.surface_tension — Method

surface_tension(w, kx)

Calculate surface tension from solution w at coordinate kx or x.

If w is dimensional provide dimensional x. If not provide k*x.

SteadyWaves.Output.vertical_velocity — Method

vertical_velocity(w, kx, kz)

Calculate vertical velocity from solution w at coordinates (kx, kz) or (x, z).

If w is dimensional provide dimensional (x, z). If not provide (k*x, k*z).

SteadyWaves.Output.wave_height — Method

wave_height(w)

Calculate wave height H from solution w.

Dimensionality of result depend on dimensionality of struct w

SteadyWaves.Output.wave_number — Method

wave_number(w)

Calculate wave number K from solution w.

Dimensionality of result depend on dimensionality of struct w

SteadyWaves.Output.wave_period — Method

wave_period(w)

Calculate wave period T from solution w.

Dimensionality of result depend on dimensionality of struct w

SteadyWaves.Output.wave_period — Method

wave_period(w::WaveStruct)

Calculate wave period T from wave struct w.

SteadyWaves.Output.wave_power — Method

wave_power(w::WaveStruct)

Calculate wave power F from wave struct w.

SteadyWaves.Output.wavelength — Method

wavelength(w)

Calculate wavelength L from solution w.

Dimensionality of result depend on dimensionality of struct w

SteadyWaves.Output.wavelength — Method

wavelength(w::WaveStruct)

Calculate wavelength L from wave struct w.

SteadyWaves.Params.CurrentCriterion — Type

Current Criterion selects condition that relates c to u

CC_EULER - euler condition u - c = 0

CC_STOKES - stokes condition u - c - Q/d = 0

CC_INVALID - for internal use only. Most likely value was:

not selected

used in invalid context

SteadyWaves.Params.DepthType — Type

DepthType selects stream function representesion

DEEP_WATER - selects representesion stable in deep water - e^y

SHALLOW_WATER - selected representesion stable in shallow water sinh(y+d)/cosh(d)

STABLE - selects representesion stable in all conditions but slow sinh(y)+tanh(d)*cosh(y)

SteadyWaves.Params.ElevationType — Type

Elevation Type selects representesion of free surface η

DIRECT_ELEVATION - represents η as N+1 points along a wave

FOURIER_ELEVATION - represents η as N+1 cosine frequencies of a wave

INVALID_ELEVATION - for internal use only. Most likely value was:

not selected

used in invalid context

SteadyWaves.Params.ParameterCriterion — Type

Parameter Criterion selects condition that enforce wave length via L or T

PC_LENGTH - length condition L - 2π = 0

PC_PERIOD - period condition T*c - 2π = 0

PC_INVALID - for internal use only. Most likely value was:

not selected

used in invalid context

SteadyWaves.Params.WaveType — Type

Wave Type selects dominant forces shaping wave

GRAVITY_WAVE - apply only gravitational force to a wave

CAPILLARY_WAVE - apply only capillary force to a wave

GRAVITYCAPILLARYWAVE - apply both capilary and gravitational forces to a wave

INVALID_WAVE - for internal use only. Most likely value was:

not selected

used in invalid context

SteadyWaves.Shoaling.topo_approx — Method

topo_approx(d, H, L; cc=2, N=10, g=G)

Calculate shoaling coefficients K in range of depth values d for wave of length L and height H.

Arguments

d: vector of decreasing water depths (m)

L: initial wavelength (m) - corresponding to d[1]

H: initial wave height (m) - corresponding to d[1]

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

g: gravity acceleration (m/s^2), defaults to g=9.81

Output

K: vector of shoaling coefficient values

SteadyWaves.Shoaling.update_depth_fourier_approx — Method

update_depth_fourier_approx(u, d, d_p, F, T; cc=1, N=10, g=G)

Update approximate solution u of a steady wave of power F and period T propagating in water of changing depth from d to d_p using Fourier Approximation Method.

Arguments

u: solution matrix being mutated

d: initial water depth (m)

d_p: target water depth (m)

F: wave power (kg m/s)

T: wave period (s)

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

SteadyWaves.Steady.fourier_approx — Method

fourier_approx(d, H, P; pc=1, cc=1, N=10, M=1, g=G)

Approximate solution u of a steady wave of height H and length L propagating in water of depth d using Fourier Approximation Method.

Arguments

d: water depth (m)

H: wave height (m)

P: wave parameter - length L (m) or period T (s)

pc: parameter criterion; pc=1, pc=PC_LENGTH - length (default), pc=2, pc=PC_PERIOD - period

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

M: number of height steps, defaults to M=1

g: gravity acceleration (m/s^2), defaults to g=9.81

Output

u[1:N+1]: free surface elevation kη

u[N+2:2N+1]: stream function coefficients B

u[2N+2]: wave celerity c√(k/g)

u[2N+3]: mean water depth kη̄

u[2N+4]: volume flux due to waves q√(k³/g)

u[2N+5]: Bernoulli constant rk/g

u[2N+6]: mean flow velocity Ū√(k/g)

u[2N+7]: wave height kH

SteadyWaves.Linear.linear_solution — Method

linear_solution(d,H,P;pc=Params.PC_LENGTH, eta_type::Params.ElevationType=Params.FOURIER_ELEVATION,wave_type::Params.WaveType=Params.GRAVITY_WAVE, g=G, rho=RHO,sigma=SIGMA,N=10)

Approximate solution u of a steady wave of height H and length L propagating in water of depth d using Linear Approximation Method.

Arguments

d: water depth (m)

H: wave height (m)

P: wave parameter - length L (m) or period T (s)

pc: parameter criterion; pc=1, pc=PC_LENGTH - length (default), pc=2, pc=PC_PERIOD - period

eta_type: elevation representation type;

wave_type: type of forces shaping the wave;

N: number of solution eigenvalues, defaults to N=10

g: gravity acceleration (m/s^2), defaults to g=9.81

rho: density (kg/m^3), defaults to ρ=1000

sigma: surface tension coefficient (kg/s^2), defaults to σ=0.073

Output

w: dimensionless wave struct

df: dimensional factor