API

SteadyWaves.SteadyWaves

SteadyWaves.Index.C_INDEX

SteadyWaves.Index.D_INDEX

SteadyWaves.Index.H_INDEX

SteadyWaves.Index.Q_INDEX

SteadyWaves.Index.R_INDEX

SteadyWaves.Index.U_INDEX

SteadyWaves.Index.eta_indexes

SteadyWaves.Index.psi_indexes

SteadyWaves.Output.indirect_wave_period

SteadyWaves.Output.indirect_wavelength

SteadyWaves.Output.wave_height

SteadyWaves.Output.wave_number

SteadyWaves.Output.wave_period

SteadyWaves.Output.wavelength

SteadyWaves.Shoaling.topo_approx

SteadyWaves.Shoaling.update_depth_fourier_approx

SteadyWaves.Steady.fourier_approx

SteadyWaves.SteadyWaves — Module

SteadyWaves is an implementation of Rienecker and Fenton (1981) Fourier Approximation Method to steady, periodic, nonlinear waves propagating in water of constant depth.

SteadyWaves.Index.C_INDEX — Constant

u[2N+C_INDEX]: wave celerity c√(k/g)

SteadyWaves.Index.D_INDEX — Constant

u[2N+D_INDEX]: mean water depth kη̄

SteadyWaves.Index.H_INDEX — Constant

u[2N+H_INDEX]: wave height kH

SteadyWaves.Index.Q_INDEX — Constant

u[2N+Q_INDEX]: volume flux due to waves q√(k³/g)

SteadyWaves.Index.R_INDEX — Constant

u[2N+R_INDEX]: Bernoulli constant rk/g

SteadyWaves.Index.U_INDEX — Constant

u[2N+U_INDEX]: mean flow velocity Ū√(k/g)

SteadyWaves.Index.eta_indexes — Method

u[eta_indexes(N)]: free surface elevations kη

SteadyWaves.Index.psi_indexes — Method

u[psi_indexes(N)] : stream function coefficients B

SteadyWaves.Output.indirect_wave_period — Method

wave_period(u, d, N; g=G)

Calculate dimensional wave period T from solution u.

SteadyWaves.Output.indirect_wavelength — Method

wavelength(u, d, N)

Calculate dimensional wavelength L from solution u.

SteadyWaves.Output.wave_height — Method

wave_height(w, df)

Calculate dimensional wave height H from solution w and dimensional factor df.

SteadyWaves.Output.wave_number — Method

wave_number(w, df)

Calculate dimensional wave number K from solution w and dimensional factor df.

SteadyWaves.Output.wave_period — Method

wave_period(w, df)

Calculate dimensional wave period T from solution w and dimensional factor df.

SteadyWaves.Output.wavelength — Method

wavelength(w, df)

Calculate dimensional wavelength L from solution w and dimensional factor df.

SteadyWaves.Shoaling.topo_approx — Method

topo_approx(d, H, L; cc=2, N=10, g=G)

Calculate shoaling coefficients K in range of depth values d for wave of length L and height H.

Arguments

d: vector of decreasing water depths (m)

L: initial wavelength (m) - corresponding to d[1]

H: initial wave height (m) - corresponding to d[1]

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

g: gravity acceleration (m/s^2), defaults to g=9.81

Output

K: vector of shoaling coefficient values

SteadyWaves.Shoaling.update_depth_fourier_approx — Method

update_depth_fourier_approx(u, d, d_p, F, T; cc=1, N=10, g=G)

Update approximate solution u of a steady wave of power F and period T propagating in water of changing depth from d to d_p using Fourier Approximation Method.

Arguments

u: solution matrix being mutated

d: initial water depth (m)

d_p: target water depth (m)

F: wave power (kg m/s)

T: wave period (s)

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

SteadyWaves.Steady.fourier_approx — Method

fourier_approx(d, H, P; pc=1, cc=1, N=10, M=1, g=G)

Approximate solution u of a steady wave of height H and length L propagating in water of depth d using Fourier Approximation Method.

Arguments

d: water depth (m)

H: wave height (m)

P: wave parameter - length L (m) or period T (s)

pc: parameter criterion; pc=1, pc=PC_LENGTH - length (default), pc=2, pc=PC_PERIOD - period

cc: current criterion; cc=1, cc=CC_STOKES - Stokes (default), cc=2, cc=CC_EULER - Euler

N: number of solution eigenvalues, defaults to N=10

M: number of height steps, defaults to M=1

g: gravity acceleration (m/s^2), defaults to g=9.81

Output

u[1:N+1]: free surface elevation kη

u[N+2:2N+1]: stream function coefficients B

u[2N+2]: wave celerity c√(k/g)

u[2N+3]: mean water depth kη̄

u[2N+4]: volume flux due to waves q√(k³/g)

u[2N+5]: Bernoulli constant rk/g

u[2N+6]: mean flow velocity Ū√(k/g)

u[2N+7]: wave height kH