API

SteadyWaves.SteadyWaves

SteadyWaves.dispertion_relation

SteadyWaves.f

SteadyWaves.fourier_approx

SteadyWaves.init_conditions

SteadyWaves.SteadyWaves — Module

SteadyWaves is an implementation of Rienecker and Fenton (1981) Fourier Approximation Method to steady, periodic, nonlinear waves propagating in water of constant depth.

SteadyWaves.dispertion_relation — Function

dispertion_relation(d, ω, g=9.81, ϵ=10^-12)

Calculate wavenumber k based on depth d, angular wave frequency ω and gravitational acceleration g for given accuracy ϵ.

SteadyWaves.f — Method

f(du, u, p)

Define nonlinear system f(u) = 0 with parameters p.

SteadyWaves.fourier_approx — Method

fourier_approx(d, H, P; pc=1, cc=1, N=10, M=1, g=9.81)

Approximate solution u of a steady wave of height H and length L propagating in water of depth d using Fourier Approximation Method.

...

Arguments

d: water depth (m)

H: wave height (m)

P: wave parameter - length L (m) or period T (s)

pc: parameter criterion; pc=1 - length (default), pc=2 - period

cc: current criterion; cc=1 - Stokes (default), cc=2 - Euler

N: number of solution eigenvalues, defaults to N=10

M: number of height steps, defaults to M=1

g: gravity acceleration (m/s^2), defaults to g=9.81

Output (nondimensional)

u[1:N+1]: free surface elevation kη

u[N+2:2N+1]: stream function coefficients B

u[2N+2]: wave celerity c√(k/g)

u[2N+3]: mean water depth kη̄

u[2N+4]: volume flux due to waves q√(k³/g)

u[2N+5]: Bernoulli constant rk/g

u[2N+6]: mean flow velocity Ū√(k/g)

SteadyWaves.init_conditions — Method

init_conditions(d, H, P, pc, N, M)

Calculate initial conditions u0 of a steady wave of height H and length L propagating in water of depth d using linear wave theory.

...

Arguments

d: water depth (m)

H: wave height (m)

P: wave parameter - length L (m) or period T (s)

pc: parameter criterion; pc=1 - length, pc=2 - period

N: number of solution eigenvalues

M: number of height steps

Output (nondimensional)

u[1:N+1]: free surface elevation kη

u[N+2:2N+1]: stream function coefficients B

u[2N+2]: wave celerity c√(k/g)

u[2N+3]: mean water depth kη̄

u[2N+4]: volume flux due to waves q√(k³/g)

u[2N+5]: Bernoulli constant rk/g

u[2N+6]: mean flow velocity Ū√(k/g)